Move the qfuction creation into
ceed-libparanumal-qfuction.c
