+++ Makefile
@@ -34,7 +34,7 @@ ifneq ($(wildcard ../mfem/libmfem.*),)
 endif
 
 # OCCA_DIR env variable should point to OCCA master (github.com/libocca/occa)
-OCCA_DIR ?= ../occa
+#OCCA_DIR ?= ../occa
 
 # env variable MAGMA_DIR can be used too
 MAGMA_DIR ?= ../magma
@@ -119,10 +119,11 @@ petscexamples.c := $(sort $(wildcard examples/petsc/*.c))
 petscexamples  := $(petscexamples.c:examples/petsc/%.c=$(OBJDIR)/petsc-%)
 
 # backends/[ref, template, blocked, occa, magma]
-ref.c      := $(sort $(wildcard backends/ref/*.c))
-template.c := $(sort $(wildcard backends/template/*.c))
-blocked.c  := $(sort $(wildcard backends/blocked/*.c))
-occa.c     := $(sort $(wildcard backends/occa/*.c))
+ref.c          := $(sort $(wildcard backends/ref/*.c))
+template.c     := $(sort $(wildcard backends/template/*.c))
+blocked.c      := $(sort $(wildcard backends/blocked/*.c))
+occa.c         := $(sort $(wildcard backends/occa/*.c))
+libparanumal.c := $(sort $(wildcard backends/libParanumal/*.c))
 magma_preprocessor := python backends/magma/gccm.py
 magma_pre_src  := $(filter-out %_tmp.c, $(wildcard backends/magma/ceed-*.c))
 magma_dsrc     := $(wildcard backends/magma/magma_d*.c)
@@ -203,6 +204,7 @@ $(libceed) : LDFLAGS += $(if $(DARWIN), -install_name @rpath/$(notdir $(libceed)
 libceed.c += $(ref.c)
 libceed.c += $(template.c)
 libceed.c += $(blocked.c)
+libceed.c += $(libparanumal.c)
 
 ifneq ($(wildcard $(OCCA_DIR)/lib/libocca.*),)
   $(libceed) : LDFLAGS += -L$(OCCA_DIR)/lib -Wl,-rpath,$(abspath $(OCCA_DIR)/lib)
